🔧 chore(package.json): add 'commit' script to automate linting, formatting, and committing changes The 'commit' script has been added to the package.json file. This script runs the 'prettier:format' script to format the code, followed by the 'lint' script to perform linting, and then adds all the changes to the git staging area using 'git add .'. Finally, it runs the 'oco' command to create a commit with a conventional commit message. This script helps automate the process of formatting, linting, and committing changes, making it easier for developers to follow the conventions and maintain code quality.
🚚 refactor(links.tsx): rename 'Link' interface to 'LinkItem' for better clarity
The 'Link' interface in the 'links.tsx' file has been renamed to 'LinkItem' to provide better clarity and improve code readability. The new name accurately reflects the purpose of the interface, which represents an item in the list of links. This change helps make the code more understandable and maintainable.

🚀 feat(links.tsx): update 'Links' component to accept 'items' and 'ParentComponent' props
The 'Links' component in the 'links.tsx' file has been updated to accept two props: 'items' and 'ParentComponent'. The 'items' prop is an array of 'LinkItem' objects, which represent the links to be rendered. The 'ParentComponent' prop is used to specify the component that will be used to render each link item. This change provides more flexibility and reusability to the 'Links' component, allowing it to be used with different components based on the requirements.

🆕 feat(navigation.tsx): create 'Navigation' component to replace 'SideNav' component
The 'Navigation' component has been created in the 'navigation.tsx' file to replace the deprecated 'SideNav' component. The new 'Navigation' component uses the '@nextui-org/react' library to render a responsive navigation bar with menu toggle functionality. It includes the 'Links' component to render the list of links. This change improves the user experience and provides a more modern and intuitive navigation interface.

🔥 remove(sidenav.tsx): remove deprecated 'SideNav' component
The deprecated 'SideNav' component has been removed from the codebase as it is no longer used. The functionality of the 'SideNav' component has been replaced by